Get basic functionality working with new Met Museum API: https://metmuseum.github.io/
The API call used to be two calls with the same shape of response. The first would be used to get the total number of responses and the second would request a subset of those items. Now the API first requires a call that just returns object ID numbers. Then we take a subset of those items and request their full details with separate API calls.

-Remove unused using statements
-Simplify some of the workarounds that were needed to pull image urls previously