Fix AWS Exception at the End of the Handler
AWS was giving this exception when finishing the handler:

Cannot access a closed Stream.: ObjectDisposedException
   at System.IO.StreamHelpers.ValidateCopyToArgs(Stream source, Stream destination, Int32 bufferSize)
   at System.IO.Stream.CopyTo(Stream destination, Int32 bufferSize)

I think this was because I was returning a stream that had already been read for input. The effect was that the exception would be thrown and the lambda would restart, posting multiple tweets. I don't want to spam the timeline so this is worth fixing.

-Change the return type of the handler to void. According to this source, since I'm only ever calling it via events, I don't need a return type https://docs.aws.amazon.com/lambda/latest/dg/dotnet-programming-model-handler-types.html