Change from using primaryImageSmall from the Met response to the primaryImage which is much larger.
I could see that the smaller image was too small for the text recognition to work, so I tried to switch to the bigger images. This made the Project Oxford image analysis fail because the images were too big. Through experimentation, I tried resizing to different sizes to see what was the largest image that would still pass through the service. It appears to have a 1 MB limit.

2056 x 2300 = 2.52 MB too large
1788 x 2000 = 1.94 MB too large
1698 x 1900 = 1.75 MB too large
1341 x 1500 = 1.1 MB too large
1251 x 1400 = 1.001 MB too large

1162 x 1300 = 0.9 MB works
1117 x 1250 = 0.8 MB works
894 x 1000 = 0.5 MB works
89 x 100 = 0.01 MB works

I've used as 1300 dimension limit to approximate this.

Add better help message display for when project oxford fails.